refactor(docs): move the model catalogue into lib
constants/ held one file, and that file is not only constants: it carries resolveModelId, getContextWindow and isAvailableModelId alongside the catalogue. lib is where this app's non-component modules already live, including lib/constants.ts.

the demo zip generates its own constants/model.ts stub for downloaded projects, so its key moves in step; the self-containment test catches the mismatch otherwise.

lib/feature-flags.ts stays where it is. it reads like a third home for constants but has thirteen consumers across the xulux routes, the playground and the learn pages, so folding it into lib/constants.ts would make all of them pull a 237 line module for one boolean.
