+++ apps/docs/components/pages/docs/assistant/footer.tsx
@@ -8,7 +8,7 @@ import { useCurrentPage } from "@/components/pages/docs/contexts/current-page";
 import { useThreadTokenUsage } from "@assistant-ui/ai-sdk";
 import { ContextDisplay } from "@assistant-ui/ui/components/assistant-ui/context-display";
 import { useSharedDocsModelSelection } from "./composer";
-import { getContextWindow } from "@/constants/model";
+import { getContextWindow } from "@/lib/model";
 
 type AssistantFooterProps = {
   onNewThread?: () => void;
